+++ src/app/navs/side-nav/side-nav.component.ts
@@ -107,11 +107,11 @@ export class SideNavComponent implements OnInit, AfterViewInit, OnDestroy {
             },
             {
               title: 'Advanced Search',
-              alias: 'reports-advanced-filters',
-              path: 'main/reports/advanced-filters',
-              parentAlias: 'reports',
+              alias: 'advanced-search',
+              path: 'main/projects/advanced-filters',
+              parentAlias: 'advSearch',
               active: false,
-              permissionProtected: false,
+              permissionProtected: true,
               hidden: false
             }
           ]