+++ src/app/projects/advanced-filters/advanced-filters.component.ts
@@ -2,10 +2,11 @@ import { Component, OnInit, OnDestroy, ViewChild, ElementRef, EventEmitter, Outp
 import { ToolsService } from '../../_shared/services/tools.service';
 import { ApiDataAdvancedFilterService, ApiDataOrgService } from '../../_shared/services/api-data/_index';
 import { Subscription } from 'rxjs/Subscription';
-import { CacheService } from '../../_shared/services/cache.service';
 import { ExcelExportService } from '../../_shared/services/excel-export.service';
 import { AdvancedFiltersTypeaheadService } from './services/advanced-filters-typeahead.service';
-import { FilterPipe } from '../../_shared/pipes/filter.pipe';
+import { AuthService } from '../../_shared/services/auth.service';
+import { CacheService } from '../../_shared/services/cache.service';
+import { Router, NavigationEnd } from '@angular/router';
 
 // import { start } from 'repl';
 const moment = require('moment');
@@ -102,14 +103,27 @@ export class AdvancedFiltersComponent implements OnInit, OnDestroy {
   fteDateFrom: any; // for FTE date range input - format yyyy-MM-dd required
   fteDateTo: any; // for FTE date range input - format yyyy-MM-dd required
 
+  // TO-DO PAUL: REMOVE TEMP CODE
+  showDashboardButton: boolean;
+
   constructor(
+    private router: Router,
     private apiDataAdvancedFilterService: ApiDataAdvancedFilterService,
     private apiDataOrgService: ApiDataOrgService,
+    private authService: AuthService,
     private cacheService: CacheService,
     private excelExportService: ExcelExportService,
     private toolsService: ToolsService,
     private advancedFiltersTypeaheadService: AdvancedFiltersTypeaheadService,
   ) {
+
+    // TO-DO PAUL: REMOVE TEMP CODE
+    if (this.authService.loggedInUser.email === 'paul_sung@keysight.com' ||
+        this.authService.loggedInUser.email === 'tawanchai.schmitz@keysight.com' ||
+        this.authService.loggedInUser.email === 'bill_schuetzle@keysight.com' ||
+        this.authService.loggedInUser.email === 'bryan.cheung@keysight.com') {
+          this.showDashboardButton = true;
+        }
 
     // declare filter option object; NULLs ar REQUIRED
     this.filterObject = {
@@ -259,8 +273,9 @@ export class AdvancedFiltersComponent implements OnInit, OnDestroy {
 
     this.showSpinner = true;
 
-    // get search results object containing flat and nested filter result objects
     const advancedFilteredResultsObj = await this.apiDataAdvancedFilterService.getAdvancedFilteredResults(filterOptions).toPromise();
+
+    this.cacheService.advancedSearchFilterOption = filterOptions;
 
     advancedFilteredResultsObj.nested.forEach( project => {
       const schedules = [];
@@ -1072,6 +1087,11 @@ export class AdvancedFiltersComponent implements OnInit, OnDestroy {
     setTimeout(() => {
       this.excelExportService.exportServer('Jarvis Projects Export', 'Projects', this.advancedFilteredResultsFlat);
     }, 1000);
+  }
+
+  onDashboardClick() {
+    // navigate to the display page
+    this.router.navigate([`/main/projects/advanced-dashboard/`]);
   }
 
   // TEST BUTTON